// src/lib/interfaces/maps/ponds.ts
export interface MotorAddData {
  id: number;
  name: string | undefined;
  power: string | undefined;
  location: [number, number];
  motorId: number;
}
export interface PondData {
  id: string;
  coordinates: number[][];
  area: number;
  name: string;
  location: any;
  motors: MotorAddData[];
}